Add RecipesController
This commit adds the RecipesController along with the /recipes/search route and rspec tests.

For simplicity sakes in this test, all of the business logic is inside the controller and there are only integration tests for it.

Usage: make a GET request to /recipes/search with comma separated ingredients as query parameters to get the relevant recipes. E.g. GET /recipes/search?egg,flour